Move ComputeIndexOwner details into separate header file
